refactor: migrate organization-related functions to user service and organization service
- Replaced direct calls to organization service with user service methods for fetching active organizations, members, and invitations.
- Updated components to utilize the new user service methods, enhancing code consistency and maintainability.
- Removed the deprecated organization service and its associated methods to streamline the service architecture.
