+++ web-app/src/app/(app)/billing/page.tsx
@@ -2,14 +2,14 @@ import BillingForm from "@/components/billing/billing-form";
 import { getEnvSecrets } from "@/config/env.secrets";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
 import { stripeClient } from "@/lib/clients/stripe.client";
-import { getActiveOrganization } from "@/lib/services";
+import { userService } from "@/lib/services";
 
 export default async function BillingPage() {
   await getSessionOrRedirect();
 
   const productId = getEnvSecrets().STRIPE_PRODUCT_ID;
   const price = await stripeClient.getPriceByProductId(productId);
-  const activeOrganization = await getActiveOrganization();
+  const activeOrganization = await userService.getActiveOrganization();
 
   return (
     <div className="mx-auto max-w-3xl space-y-6 p-6">
