+++ web-app/src/lib/actions/billing/action.ts
@@ -5,7 +5,7 @@ import { ActionError, BillingErrorCode, CommonErrorCode } from "@/lib/actions";
 import { getSession } from "@/lib/auth/utils";
 import { stripeClient } from "@/lib/clients/stripe.client";
 import { CouponError } from "@/lib/errors/coupon-errors";
-import { getMyMemberInOrganization } from "@/lib/services";
+import { userService } from "@/lib/services";
 import { stripeService } from "@/lib/services/stripe.service";
 import { Err, Ok, Result } from "@/lib/ts-res";
 
@@ -72,7 +72,8 @@ export async function purchaseCredits(
 
     // Verify user is member of the organization
     if (organizationId) {
-      const member = await getMyMemberInOrganization(organizationId);
+      const member =
+        await userService.getMyMemberInOrganization(organizationId);
       if (!member) {
         return Err({
           message: "Unauthorized",
@@ -118,7 +119,8 @@ export async function getFreeCreditsWithCoupon(
 
     // If organizationId is provided, verify user is a member
     if (organizationId) {
-      const member = await getMyMemberInOrganization(organizationId);
+      const member =
+        await userService.getMyMemberInOrganization(organizationId);
       if (!member) {
         return Err({
           message: "Unauthorized",
