+++ web-app/src/lib/services/index.ts
@@ -1,7 +1,6 @@
 export * from "./agent.service";
 export * from "./job.service";
 export * from "./lock.service";
-export * from "./organization";
 export * from "./organization.service";
 export * from "./stripe.service";
 export * from "./user.service";
