--- web-app/src/lib/services/organization/index.ts
@@ -1,1 +1,0 @@
-export * from "./service";
